requirements.txt
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-openai==0.1.8
langserve==0.2.1
langsmith==0.1.67
python-dotenv==1.0.1
streamlit==1.35.0
uvicorn==0.30.1
sse_starlette